Link group instances instead of groups. Fix broken motor link.
Also remove some 'Assembly-' prefixes because the distinction is blurry
as once an assembly is acquired as a whole part, the assembly suffix is
obsolete (and every prefix had to be removed).
Everything can be an assembly this way and finding objects in the Groups
view of the outliner is easier too.